moves create_slapp_inputs to modules
